make converter code portable
